Add tank delay; protect the building from shoots
The tank delay is under development, it does not work yet as expected.
Its goal is to increase the control of the player, because the tank
moves too fast.

The building is protected from shoots. This is an experimental feature.
Another option is to make it possible to hit the building with friend
fire.

FossilOrigin-Name: fe8813868bcc9c14c29a5b7ab2a8411afd172352b80a4e5363ca7160f75820dc