Make the big tank definitive
The 2x1 tank of the original version made it impossible to shoot from
the center of the graphic, especially when projectiles are UDG. This
caused the projectiles could fly right at the left of the building, but
not on the other side.  The 3x1 tank makes things easier, both with UDG
and pixel projectiles.

FossilOrigin-Name: 7d3e9529a3cf1d8d94fef58ac5ee5ea94d91dadf0a761171bc5bc5d4a617f759