Calculate invader's points from its altitude
So far the points were stored in the species data table.  Now the points
do not depend on the species, but on the current altitude.  The
calculation is slower than the fetching, but it's needed in order to
implement the vertical movement of the invaders.

FossilOrigin-Name: a36a8e2665bab83fe6e0ff1396b810ec9f6477fec9917d562874ee3be661e669