Use projectiles' data instead of an extra stack
Formerly an extra stack was used to store the numbers of the unused
projectiles.  Using only the data structure is an experimental method to
prepare the implementation of a maximum number of projectiles. But
searching and counting the unused projectiles is much slower this way.

FossilOrigin-Name: 548d2d55a76a284a0e6dbbfebbe68d51f67ab413c014426a39c41026f24ddbfe