Repair the tank before the attack
This fixes a bug: the timings of the tank were initiated only at the
start of the battle instead of the attack. Since the ticks clock was
initiated at the start of the attack, the tank was frozen at the start
of the second attack wave. This bug was unnoticed for some time, since
the conditions of a new attack are unusual.

FossilOrigin-Name: 8cf9025297f6a0b4cb47c25034c8ce3cf27ac2219c7786b7cc6defa471df7f73