Make balls repair the eroded bricks
Now the breaches are only the final phase of the erosion and the balls
repair one level of erosion.

The graphics of the breaches have been simplified: the breach does not
affect the bricks above and below it, which could be affected also by
nearby breaches.  This simplifies the code at the moment.

FossilOrigin-Name: 168fb74b6832a2714c0ebeaf83da8ecd73fa6bc1c01cfa06c409ebe94695ed7b